+++ python/lsst/ts/criopy/gui/data_form_widget.py
@@ -60,10 +60,10 @@ class DataFormWidget(QWidget):
         self,
         signal: Signal,
         fields: list[tuple[str | None, DataLabel]],
-        timeChart: TimeChart | None = None,
+        time_chart: TimeChart | None = None,
     ):
         super().__init__()
-        self._timeChart = timeChart
+        self.__time_chart = time_chart
 
         layout = QFormLayout()
         for text, label in fields:
@@ -77,15 +77,14 @@ class DataFormWidget(QWidget):
 
     def _process_signal(self, data: BaseMsgType) -> None:
         for e in dir(data):
-            ch = self.findChild(QWidget, e)
-            if ch is not None:
-                ch.setValue(getattr(data, e))
+            for ch in self.findChildren(QWidget, e):
+                ch.new_data(data)
 
     def mousePressEvent(self, ev: QMouseEvent) -> None:
-        if self._timeChart is not None:
+        if self.__time_chart is not None:
             child = self.childAt(ev.pos())
             if child is not None:
-                self._timeChart.topicSelected.emit(child)
+                self.__time_chart.topicSelected.emit(child)
 
 
 class DataFormButton(ColoredButton):