+++ doc/version-history.rst
@@ -4,6 +4,15 @@
 Version History
 ===============
 
+v0.8.1
+------
+
+* Correctly show small values (= calibrations)
+* Improve hardpoint display
+* Improve IMS display
+* Display outerLoop data
+* Fixed mypy issues
+
 v0.8.0
 ------